Fixed issues with unit tests and with improper refactoring of Forum.Data project
