+++ test/Forum.Tests.Library/Fixtures/DatabaseFixture.cs
@@ -9,6 +9,7 @@ public class DatabaseFixture : IDisposable
     public DatabaseFixture()
     {
         Database = new TestDatabase();
+        Database.Initialize();
     }
     
     public void Dispose()
